Slide back & Slide back land
For slide back, just like with slide forward, the beginning is new.
Note: since the engine doesn't anticipate Lara reaching the ground when sliding back, and the sliding back animation normally has one foot "in the ground", when she reaches the bottom she already has one foot in the ground for real (just for 1-2 frames but it's still visible). To avoid that I've moved the slide back animation up and forward (64 units). The only side effect is that when she starts sliding she's a little bit higher on the slope, but it's less a problem than her ending inside the ground.