Small forward jump fix (slopes)
Fixed jumping forward above a 4-clicks high slope:
Now Lara is able to *get on the floor* ♪(^∇^*) on top of the slope if you hold the Up key, like in the classics. This might also fix a subtle visual issue at the beginning of the jump.
It *might* also break some things because the first frame of the jump animation itself is skipped, but theoretically it should be fine... or I'm really unlucky.