Double Doors
Yes, I know there's a gap between Lara's hands and the door (64 units), but it's not even noticeable in most situations (when the camera is behind Lara).

At this point I just focus on making the animations work, and then I wanna get correct positions (and behaviours) in TR5Main. I'm tired of all the hardcoded stuff.