Small jump (walk+jump) & fix for landing after sprint jump
Small jump is just like in AOD: hold Walk and press and hold Jump.
Swan diving is still possible, just hold Up as well.
But the implementation might need to be tuned cause it's easy to fail at making that new small jump.

Sprint jumping is yet to be remade - but now she won't have that glitch when she lands (the beginning of the animation was cutoff)