Splat & Push object & Lara sounds small update
Splat animation is the one when you hit a wall while jumping. In AOD it plays *sometimes*, but most of the time it's cutoff for some reason.

Push object is used in AOD when pushing the statue in St Aicard's graveyard.
It's now used for the big wall switches (sequence switches, used in TR4 in Underneath The Sphinx) and pushing rolling balls.

Lara's sounds have been extracted from AOD directly again, and I've studied how they were used in AOD to replicate it more closely.