WIP "medium hit back" animation with recover
Press Numpad2 to test it: Lara will fall as if she was hit from behind. If you press Jump quick enough, she recovers with a roll.
It's planned to have a transition into running, it's planned to have a proper animation when falling above a ledge, etc... and have animations for each side (for now, AOD's "big hit front" animation is used for a hit from the front, press Numpad8 to test it)... it's planned whatever is needed to make combat more fun.
But as always I have my doubts about TRNG's capabilities and if it's really possible to make this work correctly (for example, Raildex couldn't properly force Lara to not use her weapons during his hit animations). We'll see how it goes, and if it's not for TRNG it will be for TEN anyway!